delete target button

+++ src/main/java/ddns/net/data/repository/TargetRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 public interface TargetRepository extends CrudRepository<Target,Integer> {
 
     Target findOneByName(String name);
-    Target findOneById(int id);
+    Target findOneById(long id);
     Target findOneByEmail(String email);
 
 }
