+++ src/main/java/ddns/net/data/service/LocationDataServiceImpl.java
@@ -35,7 +35,7 @@ public class LocationDataServiceImpl implements LocationDataService {
 
     @Transactional(readOnly = true)
     @Override
-    public List<LocationData> findAllByChildId(int id) {
+    public List<LocationData> findAllByChildId(long id) {
         return locationDataRepository.findAllByTargetId(id);
     }
