+++ src/main/java/ddns/net/data/service/TargetService.java
@@ -6,7 +6,7 @@ public interface TargetService {
 
     Target save(Target child);
     Target findOneByName(String name);
-    Target findOneById(int id);
+    Target findOneById(long id);
     Target findOneByEmail(String email);
 
 }
